Updated test harness - int macros, strings, linked libraries, debugging

+++ include/test/test.h
@@ -5,35 +5,35 @@
 #include <avr/io.h>
 #include <stdint.h>
 
-#define ASSERT_EQ(a, b) (print("AS EQ %d %d (%s) (%d)\r\n",\
-    (a), (b), __FUNCTION__, __LINE__))
+#define ASSERT_EQ(a, b) (print("AS EQ %ld %ld (%s) (%d)\r\n",\
+    (int32_t)(a), (int32_t)(b), __FUNCTION__, (int16_t)__LINE__))
 
-#define ASSERT_NEQ(a, b) (print("AS NEQ %d %d (%s) (%d)\r\n",\
-    (a), (b), __FUNCTION__, __LINE__))
+#define ASSERT_NEQ(a, b) (print("AS NEQ %ld %ld (%s) (%d)\r\n",\
+    (int32_t)(a), (int32_t)(b), __FUNCTION__, (int16_t)__LINE__))
 
-#define ASSERT_GREATER(a, b) (print("AS GT %d %d (%s) (%d)\r\n",\
-    (a), (b), __FUNCTION__, __LINE__))
+#define ASSERT_GREATER(a, b) (print("AS GT %ld %ld (%s) (%d)\r\n",\
+    (int32_t)(a), (int32_t)(b), __FUNCTION__, (int16_t)__LINE__))
 
-#define ASSERT_LESS(a, b) (print("AS LT %d %d (%s) (%d)\r\n",\
-    (a), (b), __FUNCTION__, __LINE__))
+#define ASSERT_LESS(a, b) (print("AS LT %ld %ld (%s) (%d)\r\n",\
+    (int32_t)(a), (int32_t)(b), __FUNCTION__, (int16_t)__LINE__))
 
-#define ASSERT_TRUE(v) (print("AS TRUE %d (%s) (%d)\r\n",\
-    (v), __FUNCTION__, __LINE__))
+#define ASSERT_TRUE(v) (print("AS TRUE %ld (%s) (%d)\r\n",\
+    (int32_t)(v), __FUNCTION__, (int16_t)__LINE__))
 
-#define ASSERT_FALSE(v) (print("AS FALSE %d (%s) (%d)\r\n",\
-    (v), __FUNCTION__, __LINE__))
+#define ASSERT_FALSE(v) (print("AS FALSE %ld (%s) (%d)\r\n",\
+    (int32_t)(v), __FUNCTION__, (int16_t)__LINE__))
 
 #define ASSERT_FP_EQ(a, b) (print("AS FP EQ %.3f %.3f (%s) (%d)\r\n",\
-    (float)(a), (float)(b), __FUNCTION__, __LINE__))
+    (float)(a), (float)(b), __FUNCTION__, (int16_t)__LINE__))
 
 #define ASSERT_FP_NEQ(a, b) (print("AS FP NEQ %.3f %.3f (%s) (%d)\r\n",\
-    (float)(a), (float)(b), __FUNCTION__, __LINE__))
+    (float)(a), (float)(b), __FUNCTION__, (int16_t)__LINE__))
 
 #define ASSERT_FP_GREATER(a, b) (print("AS FP GT %.3f %.3f (%s) (%d)\r\n",\
-    (float)(a), (float)(b), __FUNCTION__, __LINE__))
+    (float)(a), (float)(b), __FUNCTION__, (int16_t)__LINE__))
 
 #define ASSERT_FP_LESS(a, b) (print("AS FP LT %.3f %.3f (%s) (%d)\r\n",\
-    (float)(a), (float)(b), __FUNCTION__, __LINE__))
+    (float)(a), (float)(b), __FUNCTION__, (int16_t)__LINE__))
 
 typedef void(*test_fn_t)(void);
